requirements.txt
Django>=1.10
django-braces>=1.9,<2.0
django-extensions>=1.7
django-filter>=0.14.0,<1.0
django-nose>=1.4
django-reversion>=2.0,<2.1
djangorestframework>=3.4
factory_boy>=2.7
nose>=1.3
